2021 Day 1: Count depth measurements
Reads in a file of depth measurements and counts when the
depth increases. There are two techniques for doing this. One
is to count the increase of each individual measure and the
second is to use a sliding window, where we add 3
measurements together and and compare them to the sum of the
next 3 measurements.